+++ src/arch/aarch64/kernel/mod.rs
@@ -124,7 +124,7 @@ pub fn boot_next_processor() {
 
 		use memory_addresses::VirtAddr;
 
-		use crate::kernel::start::{TTBR0, smp_start};
+		use crate::arch::aarch64::kernel::start::{TTBR0, smp_start};
 		use crate::mm::virtual_to_physical;
 
 		if cpu_online == 0 {
