+++ src/arch/x86_64/kernel/kernel_stack.rs
@@ -1,6 +1,6 @@
 use core::mem;
 
-use crate::core_local::CoreLocal;
+use crate::arch::kernel::core_local::CoreLocal;
 
 type Reg = mem::MaybeUninit<usize>;
